requirements.txt
bleach 
Flask==1.1.2
gevent 
glob2==0.7
greenlet
gunicorn
h5py==3.2.1
HeapDict==1.0.1
html5lib 
intervaltree 
jdcal==1.4.1
jedi 
Jinja2==2.11.2
keras
matplotlib 
numpy 
numpydoc 
olefile==0.46
openpyxl 
packaging==20.4
pandas
seaborn 
statsmodels==0.12.0
tensorflow-cpu
unicodecsv==0.14.1
webencodings==0.5.1
wtforms